split client demo: app code vs SDK machinery
sdkLib.ts (~110 lines): type shims, IncompleteResult parsing, the
retry loop, withMrtr() helper. Stand-in for what the SDK ships.

clientDualPath.ts (~55 lines): just the app-developer surface.
handleElicitation + one withMrtr() call + one callTool() call.
The point being: the app code is identical to today.
